load by template

+++ automated/load_virtuoso-switch.sh
@@ -15,7 +15,7 @@ VIRTUOSO_DIR="${DOCKER_ROOT_DIR}/data"				# rundeckコンテナでのvirtuosoフ
 DOCKER_DIR="${DOCKER_ROOT_DIR}/virtuoso-switch"			# virtuosoのDockerfile配置先
 DOCKER_LOG_DIR="${DOCKER_ROOT_DIR}/virtuoso-switch_logs"	# ログの出力先
 VIRTUOSO_HOST_DIR="${ROOT_DIR}/data"				# virtuosoファイル出力先 
-TEMPLATE_VIRTUOSO_DIR=""					# gnomad当の事前ファイルがロードされているvirtuosoファイル
+TEMPLATE_VIRTUOSO_DIR="/mnt/share/dsatoh/togovar/2021.1/virtuoso/template"	# gnomad当の事前ファイルがロードされているvirtuosoファイル
 
 # job2.lckが存在する場合、異常終了
 if [ -e ${VIRTUOSO_DIR}/job2.lck ]; then 
@@ -37,28 +37,27 @@ echo "copy file"
 
 # コピー前のハッシュ値取得
 #echo `md5sum ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.trx | awk '{ print $1 }'` > ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.trx_md5
-#echo `md5sum ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.db | awk '{ print $1 }'` > ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.db_md5
-#
+time echo `md5sum ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.db | awk '{ print $1 }'` > ${VIRTUOSO_DIR}/virtuoso.db_before_md5
+
 # テンプレートのコピー
 # cp ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.trx ${VIRTUOSO_DIR}/virtuoso.trx  
-# cp ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.db ${VIRTUOSO_DIR}/virtuoso.db 
-#
-#
+time cp ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.db ${VIRTUOSO_DIR}/virtuoso.db 
+
 # コピー後のハッシュ値取得
 #echo `md5sum ${VIRTUOSO_DIR}/virtuoso.trx | awk '{ print $1 }'` > ${VIRTUOSO_DIR}/virtuoso.trx_md5
-#echo `md5sum ${VIRTUOSO_DIR}/virtuoso.db | awk '{ print $1 }'` > ${VIRTUOSO_DIR}/virtuoso.db_md5
-#
+time echo `md5sum ${VIRTUOSO_DIR}/virtuoso.db | awk '{ print $1 }'` > ${VIRTUOSO_DIR}/virtuoso.db_after_md5
+
 # コピー前後のMDハッシュ値比較
 #diff ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.trx_md5 ${VIRTUOSO_DIR}/virtuoso.trx_md5
 #exec_1=`echo $?`
-#diff ${TEMPLATE_VIRTUOSO_DIR}/virtuoso.db_md5 ${VIRTUOSO_DIR}/virtuoso.db_md5
-#exec_2=`echo $?`
-#
+diff ${VIRTUOSO_DIR}/virtuoso.db_before_md5 ${VIRTUOSO_DIR}/virtuoso.db_after_md5
+exec_2=`echo $?`
+
 # コピーしたファイルのハッシュ値が一致しない場合異常終了する
-#if [ ${exec_1} -ne 0 ] || [ ${exec_2} -ne 0 ]; then
-#  echo "DBファイルのコピーに失敗しました"
-#  exit 1
-#fi
+if [ ${exec_2} -ne 0 ]; then
+  echo "DBファイルのコピーに失敗しました"
+  exit 1
+fi
 
 # Docker の更新
 docker rmi virtuoso-switch
@@ -74,6 +73,7 @@ echo "load finish"
 # ロード結果確認、エラーがあれば出力して異常終了
 if [ -s ${DOCKER_LOG_DIR}/${YYYYMMDD}_stderr.log ]; then
   cat ${DOCKER_LOG_DIR}/${YYYYMMDD}_stderr.log >&2
+  rm ${VIRTUOSO_DIR}/job2.lck
   exit 1
 fi
